backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.2
pydantic==2.9.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.12
gunicorn==23.0.0
